backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
python-dotenv==1.0.0
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
